#31: PlaceMapActivity.kt 파일의 지도 구현을 kakaoMap에서 googleMap으로 변경
